+++ src/utils.ts
@@ -250,8 +250,9 @@ export function isPlaygroundGist(gist: Gist) {
     gistFiles.includes(PLAYGROUND_FILE) ||
     gistFiles.includes("index.html") ||
     gistFiles.includes("index.pug") ||
-    (gistFiles.includes("scripts") &&
-      gistFiles.some((file) => path.extname(file) === ".markdown"))
+    gistFiles.includes("scripts") ||
+      (gistFiles.includes("script.js") &&
+        gistFiles.some((file) => path.extname(file) === ".markdown"))
   );
 }